Remove the custom typing system of casacore
Casacore was using a manual system to identify runtime type information,
using many virtual type() methods and Register() overloads taking
dummy variables. All this functionality was easy to replace with
'normal' run-time type info using dynamic_cast or typeid(), saving
a lot of code, and improving performance in several places.

Removing these type() and variant methods highlighted several bugs
with measures like mDirection where it was
supposed to call mDirection.getRefPtr()->getType()
(which returns an enum describing something like ITRF), but
instead called mDirection.type(), returning an int that
was assigned to the MDirection type by the custom type system.

Many classes still have an 'assure()' method that is supposed
to throw an exception when the object is not of a specified
type. These are now useless, as calls can just perform
dynamic_cast<T&>, which throws when the type is not of the
specified type. However, I left those here for now.